+++ sdk/attestation/azure-security-attestation/src/main/java/com/azure/security/attestation/Utilities.java
@@ -5,11 +5,23 @@ package com.azure.security.attestation;
 
 import com.azure.core.http.rest.Response;
 import com.azure.core.http.rest.ResponseBase;
+import com.azure.core.util.logging.ClientLogger;
+import com.azure.security.attestation.implementation.models.JsonWebKey;
+import com.azure.security.attestation.implementation.models.JsonWebKeySet;
+import com.azure.security.attestation.models.AttestationSigner;
+
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.security.cert.Certificate;
+import java.security.cert.CertificateException;
+import java.security.cert.CertificateFactory;
+import java.security.cert.X509Certificate;
+import java.util.Base64;
 
 /**
  * Utility class with helper functions.
  */
-public class Utilities {
+class Utilities {
 
     /**
      * Generates a new public response type from an internal model type.
@@ -26,4 +38,59 @@ public class Utilities {
             value,
             null);
     }
+
+    /**
+     * Convert a base64 encoded string into a byte stream.
+     * @param base64 - Base64 encoded string to be decoded
+     * @return stream of bytes encoded in the base64 encoded string.
+     */
+    static InputStream base64ToStream(String base64) {
+        byte[] decoded = Base64.getDecoder().decode(base64);
+        return new ByteArrayInputStream(decoded);
+    }
+
+    /**
+     * Private method to create an AttestationSigner from a JWKS.
+     * @param jwks JWKS to create.
+     * @return Array of {@link AttestationSigner}s created from the JWK.
+     */
+    static AttestationSigner[] attestationSignersFromJwks(JsonWebKeySet jwks) {
+        return jwks
+            .getKeys()
+            .stream()
+            .map(Utilities::attestationSignerFromJwk)
+            .toArray(AttestationSigner[]::new);
+    }
+
+
+    /**
+     * Private method to create an AttestationSigner from a JWK.
+     * @param jwk JWK to create.
+     * @return {@link AttestationSigner} created from the JWK.
+     */
+    static AttestationSigner attestationSignerFromJwk(JsonWebKey jwk) {
+        ClientLogger logger = new ClientLogger(Utilities.class);
+
+        CertificateFactory cf;
+        try {
+            cf = CertificateFactory.getInstance("X.509");
+        } catch (CertificateException e) {
+            throw logger.logExceptionAsError(new RuntimeException(e.getMessage()));
+        }
+        X509Certificate[] certificates = jwk.getX5C().stream().map(base64cert -> {
+            Certificate cert = null;
+            try {
+                cert = cf.generateCertificate(Utilities.base64ToStream(base64cert));
+            } catch (CertificateException e) {
+                throw logger.logExceptionAsError(new RuntimeException(e.getMessage()));
+            }
+
+            return (X509Certificate) cert;
+        }).toArray(X509Certificate[]::new);
+
+        return new AttestationSigner()
+            .keyId(jwk.getKid())
+            .certificates(certificates);
+    }
+
 }